refactor(orulo): extract images payloads processor
